requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
chromadb
pytest
